requirements.txt
numpy~=1.23.3
python-chess
pytz~=2022.4
humanize
tensorboard
tensorflow~=2.10.0
protobuf
tensorboardx
seaborn~=0.12.0
matplotlib~=3.6.0
selenium~=4.5.0
chess~=1.9.3
PyYAML~=6.0
pandas~=1.5.0
requests~=2.28.1
websockets~=10.3
opencv-python
Pillow
undetected-chromedriver